mqqtToMongoDB/requirements.txt
DateTime==4.3
paho-mqtt==1.4.0
pymongo==3.10.1
requests==2.22.0
